Move typeorm connection options to their own file.
