--- src/ormconfig.json
@@ -1,14 +1,0 @@
-{
-  "type": "sqlite",
-  "database": "./database.sqlite",
-  "synchronize": true,
-  "logging": false,
-  "entities": ["./entity/**/*.ts"],
-  "migrations": ["./migration/**/*.ts"],
-  "subscribers": ["./subscriber/**/*.ts"],
-  "cli": {
-    "entitiesDir": "./entity",
-    "migrationsDir": "./migration",
-    "subscribersDir": "./subscriber"
-  }
-}
