+++ src/package.json
@@ -11,7 +11,8 @@
   },
   "scripts": {
     "electron-rebuild": "node -r ../.erb/scripts/BabelRegister.js ../.erb/scripts/ElectronRebuild.js",
-    "postinstall": "yarn electron-rebuild"
+    "postinstall": "yarn electron-rebuild",
+    "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js"
   },
   "license": "MIT",
   "dependencies": {
@@ -19,7 +20,5 @@
     "sqlite3": "^5.0.2",
     "typeorm": "^0.2.31"
   },
-  "devDependencies": {
-    "@types/node": "^14.14.31"
-  }
+  "devDependencies": {}
 }
